Hoist tupl_platform, change tupl_impl.hpp search
tupl_platform.hpp header is moved to top, as it will be in
tupl_amalgam.hpp, cat'd in straight after the banner.

tupl_impl.hpp search order now looks for unprefixed first
then tupl_dev/tupl_impl.hpp which is a defined target.
This allows the user to override the tupl_dev version
by placing an unqualified version in the include path.